Move UI helper funcs to their own file
